refactor: split data/fs/project_resolver.py into focused modules
